Crate seax_util

Seax Utilities

Utility library containing code for building programs involving the Seax platform. Seax is a virtual-machine-based runtime environment for functional programming languages.

This crate contains the following modules:

cell: Contains the definitions of all Seax Virtual Machine cell types, including instruction, atom, and list cells.

list: Contains the singly-linked list and stack implementations used by the Seax VM internally.

bytecode: Contains functions for encoding and decoding Seax VM cells to and from Seax bytecode

compiler_tools: Contains reusable code for implementing compilers targeting the Seax platform, including traits for abstract syntax trees and symbol tables.

Reexports

pub use self::list::{List, Stack};

pub use self::list::List::{Cons, Nil};

pub use self::cell::{SVMCell, Atom, Inst};

Modules

bytecode
Bytecode

cell
Cell

compiler_tools
Compiler Tools

list
List

Macros

list!
Convenience macro for making lists.

list_cell!