Module seax_util::compiler_tools

Compiler Tools

Contains general-purpose code for compilers targeting the Seax platform.

Reexports

pub use self::forktable::ForkTable;

pub use self::ast::ASTNode;

Modules

ast
AST

forktable
ForkTable

Traits

Scope
Trait for a symbol table

Type Definitions

CompileResult
A CompileResult is either Ok(SVMCell) or Err(&str)

Index
Represents an index into the Seax VM's environment stack.

SymTable
The symbol table for bound names is represented as a ForkTable mapping &str (names) to (uint,uint) tuples, representing the location in the $e stack storing the value bound to that name.