Trait seax_util::compiler_tools::Scope

pub trait Scope<K> where K: Eq + Hash {
    fn bind(&mut self, name: K, lvl: u64) -> Index;
    fn lookup(&self, name: &K) -> Option<Index>;
}

Trait for a symbol table

Required Methods

fn bind(&mut self, name: K, lvl: u64) -> Index

Bind a name to a scope.

Returnsthe indices for that name in the SVM environment.

fn lookup(&self, name: &K) -> Option<Index>

Look up a name against a scope.

Returns the indices for that name in the SVM environment, or None if that name is unbound.