Module seax_util::list

List

Contains singly-linked list and stack implementations.

List<T> is a singly-linked cons list. Stack<T> is a trait providing stack operations(push(), pop(), and peek()), and an implementation for List.

Reexports

pub use list::List::{Cons, Nil};

Structs

ListIterator
Wraps a List to allow it to be used as an Iterator

Enums

List
Singly-linked cons list.

Traits

Stack
Common functions for an immutable Stack abstract data type.